requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
pillow>=9.5.0
thop>=0.1.0
tqdm>=4.65.0
tensorboard>=2.13.0
scipy>=1.10.0
pycocotools>=2.0.6
opencv-python>=4.8.0
scikit-image>=0.21.0

